requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
Pillow>=10.0.0
python-dotenv==1.0.1
reportlab>=4.0.0
